Change package names in StandAloneExamples to io.pravega... from com.emc.pravega...
Fix one small naming bug in SharedConfigCli help
