+++ standalone-examples/bin/.gitignore
@@ -1,1 +1,2 @@
 /com/
+/io/
