+++ standalone-examples/src/main/java/io/pravega/example/consolerw/ConsoleReader.java
@@ -3,7 +3,7 @@
  *  Copyright (c) 2017 Dell Inc., or its subsidiaries.
  *
  */
-package com.emc.pravega.example.gettingstarted.hello;
+package io.pravega.example.consolerw;
 
 import java.net.URI;
 import java.util.Collections;
@@ -29,34 +29,38 @@ import com.emc.pravega.stream.Sequence;
 import com.emc.pravega.stream.StreamConfiguration;
 import com.emc.pravega.stream.impl.JavaSerializer;
 
-/**
- * A simple example app that uses a Pravega Reader to read from a given scope and stream.
- */
-public class HelloWorldReader {
-    private static final int READER_TIMEOUT_MS = 2000;
+import io.pravega.example.gettingstarted.HelloWorldReader;
 
+/**
+ * Reads from a Stream until interrupted.
+ */
+public class ConsoleReader {
+    
+    private static final int READER_TIMEOUT_MS = 200000;
+    
     public final String scope;
     public final String streamName;
     public final URI controllerURI;
 
-    public HelloWorldReader(String scope, String streamName, URI controllerURI) {
+    public ConsoleReader(String scope, String streamName, URI controllerURI) {
         this.scope = scope;
         this.streamName = streamName;
         this.controllerURI = controllerURI;
     }
-
+    
     public void run() {
+        final String readerGroup = UUID.randomUUID().toString().replace("-", "");
+        final ReaderGroupConfig readerGroupConfig = ReaderGroupConfig.builder().startingPosition(Sequence.MIN_VALUE)
+                .build();
         StreamManager streamManager = StreamManager.create(controllerURI);
-        
-        final boolean scopeIsNew = streamManager.createScope(scope);
-        StreamConfiguration streamConfig = StreamConfiguration.builder()
+        streamManager.createScope(scope);
+
+        StreamConfiguration streamConfig = StreamConfiguration.builder().scope(scope).streamName(streamName)
                 .scalingPolicy(ScalingPolicy.fixed(1))
                 .build();
-        final boolean streamIsNew = streamManager.createStream(scope, streamName, streamConfig);
 
-        final String readerGroup = UUID.randomUUID().toString().replace("-", "");
-        final ReaderGroupConfig readerGroupConfig = ReaderGroupConfig.builder()
-                .build();
+        streamManager.createStream(scope, streamName, streamConfig);
+
         try (ReaderGroupManager readerGroupManager = ReaderGroupManager.withScope(scope, controllerURI)) {
             readerGroupManager.createReaderGroup(readerGroup, readerGroupConfig, Collections.singleton(streamName));
         }
@@ -66,23 +70,23 @@ public class HelloWorldReader {
                                                                            readerGroup,
                                                                            new JavaSerializer<String>(),
                                                                            ReaderConfig.builder().build())) {
-            System.out.format("Reading all the events from %s/%s%n", scope, streamName);
+            System.out.format("******** Reading events from %s/%s%n", scope, streamName);
             EventRead<String> event = null;
             do {
                 try {
                     event = reader.readNextEvent(READER_TIMEOUT_MS);
-                    if (event.getEvent() != null) {
-                        System.out.format("Read event '%s'%n", event.getEvent());
+                    if(event != null) {
+                        System.out.format("'%s'%n", event.getEvent());
                     }
                 } catch (ReinitializationRequiredException e) {
                     //There are certain circumstances where the reader needs to be reinitialized
                     e.printStackTrace();
                 }
-            } while (event.getEvent() != null);
-            System.out.format("No more events from %s/%s%n", scope, streamName);
+            }while(true);
         }
     }
 
+    
     public static void main(String[] args) {
         Options options = getOptions();
         CommandLine cmd = null;
@@ -91,7 +95,7 @@ public class HelloWorldReader {
         } catch (ParseException e) {
             e.printStackTrace();
             final HelpFormatter formatter = new HelpFormatter();
-            formatter.printHelp("HelloWorldReader", options);
+            formatter.printHelp("ConsoleReader", options);
         }
         
         final String scope = cmd.getOptionValue("scope") == null ? Constants.DEFAULT_SCOPE : cmd.getOptionValue("scope");
@@ -99,8 +103,8 @@ public class HelloWorldReader {
         final String uriString = cmd.getOptionValue("uri") == null ? Constants.DEFAULT_CONTROLLER_URI : cmd.getOptionValue("uri");
         final URI controllerURI = URI.create(uriString);
         
-        HelloWorldReader hwr = new HelloWorldReader(scope, streamName, controllerURI);
-        hwr.run();
+        ConsoleReader reader = new ConsoleReader(scope, streamName, controllerURI);
+        reader.run();
     }
 
     private static Options getOptions() {
@@ -116,4 +120,5 @@ public class HelloWorldReader {
         CommandLine cmd = parser.parse(options, args);
         return cmd;
     }
+
 }
