+++ standalone-examples/src/main/java/io/pravega/example/gettingstarted/Constants.java
@@ -3,7 +3,7 @@
  *  Copyright (c) 2017 Dell Inc., or its subsidiaries.
  *
  */
-package com.emc.pravega.example.consolerw;
+package io.pravega.example.gettingstarted;
 
 /**
  * Defines a handful of constants shared by classes in this package.
@@ -11,6 +11,9 @@ package com.emc.pravega.example.consolerw;
  */
 public class Constants {
     protected static final String DEFAULT_SCOPE = "examples";
-    protected static final String DEFAULT_STREAM_NAME = "someStream";
+    protected static final String DEFAULT_STREAM_NAME = "helloStream";
     protected static final String DEFAULT_CONTROLLER_URI = "tcp://127.0.0.1:9090";
+    
+    protected static final String DEFAULT_ROUTING_KEY = "helloRoutingKey";
+    protected static final String DEFAULT_MESSAGE = "hello world";
 }
