+++ standalone-examples/src/main/java/io/pravega/example/gettingstarted/HelloWorldWriter.java
@@ -3,7 +3,7 @@
  *  Copyright (c) 2017 Dell Inc., or its subsidiaries.
  *
  */
-package com.emc.pravega.example.gettingstarted.hello;
+package io.pravega.example.gettingstarted;
 
 import java.net.URI;
